refactor: create use-deep-link-options.ts
